limited to two airports since this is about forms and all these airports was making this into a database exercise

+++ db/migrate/20210315231638_remove_departure_time_from_flight.rb
@@ -1,5 +1,5 @@
 class RemoveDepartureTimeFromFlight < ActiveRecord::Migration[6.0]
   def change
-    remove_column :flights, :departure_time, :decimal
+    remove_column :flights, :departure_time, :integer
   end
 end
